+++ .github/scripts/release-workflow-package-push.sh
@@ -11,6 +11,7 @@ while IFS="" read -r -d "" buildpack_toml_path; do
 	buildpack_version="$(yj -t <"${buildpack_toml_path}" | jq -r .buildpack.version)"
 	buildpack_docker_repository="$(yj -t <"${buildpack_toml_path}" | jq -r .metadata.release.docker.repository)"
 	buildpack_path=$(dirname "${buildpack_toml_path}")
+	buildpack_build_path="${buildpack_path}"
 
 	if [[ $buildpack_id == "${REQUESTED_BUILDPACK_ID}" ]]; then
 		# Some buildpacks require a build step before packaging. If we detect a build.sh script, we execute it and
@@ -25,7 +26,7 @@ while IFS="" read -r -d "" buildpack_toml_path; do
 		fi
 
 		image_name="${buildpack_docker_repository}:${buildpack_version}"
-		pack package-buildpack --config "${buildpack_path}/package.toml" --publish "${image_name}"
+		pack buildpack package --config "${buildpack_build_path}/package.toml" --publish "${image_name}"
 
 		# We might have local changes after building and/or shimming the buildpack. To ensure scripts down the pipeline
 		# work with a clean state, we reset all local changes here.