+++ buildpacks/nodejs/CHANGELOG.md
@@ -3,6 +3,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Add license to buildpack.toml ([#17](https://github.com/heroku/buildpacks-node/pull/17))
 
 ## [0.7.1] 2021/01/20
 - Replace logging style to match style guides ([#63](https://github.com/heroku/nodejs-engine-buildpack/pull/63))